requirements.txt
telethon
openai
matplotlib
pandas
tqdm
python-dotenv 